+++ trunk/bbs/utility.h
@@ -35,7 +35,7 @@ double ratio();
 double post_ratio();
 double nsl();
 void Wait(double d);
-double freek1(const char *pszPathName);
+long freek1(const char *pszPathName);
 void send_net(net_header_rec * nh, unsigned short int *list, const char *text, const char *byname);
 void giveup_timeslice();
 char *stripfn(const char *pszFileName);
